+++ apps/api/src/app/user/user.service.ts
@@ -13,8 +13,10 @@ import { EconomicMarketClusterRiskDevelopedMarkets } from '@ghostfolio/api/model
 import { EconomicMarketClusterRiskEmergingMarkets } from '@ghostfolio/api/models/rules/economic-market-cluster-risk/emerging-markets';
 import { EmergencyFundSetup } from '@ghostfolio/api/models/rules/emergency-fund/emergency-fund-setup';
 import { FeeRatioInitialInvestment } from '@ghostfolio/api/models/rules/fees/fee-ratio-initial-investment';
+import { RegionalMarketClusterRiskAsiaPacific } from '@ghostfolio/api/models/rules/regional-market-cluster-risk/asia-pacific';
 import { RegionalMarketClusterRiskEmergingMarkets } from '@ghostfolio/api/models/rules/regional-market-cluster-risk/emerging-markets';
 import { RegionalMarketClusterRiskEurope } from '@ghostfolio/api/models/rules/regional-market-cluster-risk/europe';
+import { RegionalMarketClusterRiskJapan } from '@ghostfolio/api/models/rules/regional-market-cluster-risk/japan';
 import { RegionalMarketClusterRiskNorthAmerica } from '@ghostfolio/api/models/rules/regional-market-cluster-risk/north-america';
 import { ConfigurationService } from '@ghostfolio/api/services/configuration/configuration.service';
 import { I18nService } from '@ghostfolio/api/services/i18n/i18n.service';
@@ -272,6 +274,12 @@ export class UserService {
         undefined,
         undefined
       ).getSettings(user.Settings.settings),
+      RegionalMarketClusterRiskAsiaPacific:
+        new RegionalMarketClusterRiskAsiaPacific(
+          undefined,
+          undefined,
+          undefined
+        ).getSettings(user.Settings.settings),
       RegionalMarketClusterRiskEmergingMarkets:
         new RegionalMarketClusterRiskEmergingMarkets(
           undefined,
@@ -279,6 +287,11 @@ export class UserService {
           undefined
         ).getSettings(user.Settings.settings),
       RegionalMarketClusterRiskEurope: new RegionalMarketClusterRiskEurope(
+        undefined,
+        undefined,
+        undefined
+      ).getSettings(user.Settings.settings),
+      RegionalMarketClusterRiskJapan: new RegionalMarketClusterRiskJapan(
         undefined,
         undefined,
         undefined
@@ -333,7 +346,10 @@ export class UserService {
           currentPermissions,
           permissions.accessHoldingsChart,
           permissions.createAccess,
-          permissions.readAiPrompt
+          permissions.createMarketDataOfOwnAssetProfile,
+          permissions.readAiPrompt,
+          permissions.readMarketDataOfOwnAssetProfile,
+          permissions.updateMarketDataOfOwnAssetProfile
         );
 
         // Reset benchmark